+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fix
+
+- Seperate refund and changeEvent badge from success(web)/admin
+- No system payment change after update
+- Stat of ubication sell_type
+- TS and HTML warnings
+
 ## [0.0.27] - 2026-05-01
 
 ### Chore